Séparation des fonctions dans un module et MAJ du readme
